+++ lib/src/state/StateValue.dart
@@ -1,0 +1,12 @@
+/// A data object that holds a retrieved state value with its key.
+class StateValue<T> {
+  /// A unique state key
+  String key;
+
+  /// A stored state value;
+  T? value;
+
+  StateValue(String key, T? value)
+      : key = key,
+        value = value;
+}